fix: account for clean lib/ dir while syncing lockfile
